test(allocation): split ListAllocationsControllerTest by behaviour
Split pagination, filters, and hospital-scope coverage into focused
classes; share setup via TestCase with Access and FilterEdges.
